move call_methods
